+++ runtime/Cpp/runtime/src/ConsoleErrorListener.cpp
@@ -36,6 +36,6 @@ using namespace antlr4;
 ConsoleErrorListener ConsoleErrorListener::INSTANCE;
 
 void ConsoleErrorListener::syntaxError(IRecognizer * /*recognizer*/, Token * /*offendingSymbol*/,
-  size_t line, int charPositionInLine, const std::string &msg, std::exception_ptr /*e*/)  {
+  size_t line, size_t charPositionInLine, const std::string &msg, std::exception_ptr /*e*/)  {
   std::cerr << "line " << line << ":" << charPositionInLine << " " << msg << std::endl;
 }
